@@ -3,27 +3,25 @@
     "displayName": "Fleece",
     "description": "Fleece is a natural language code generator for Visual Studio Code, powered by Llama and Alpaca language models.",
     "version": "0.0.1",
-    "publisher": "kelden",
+    "publisher": "cafesao",
     "repository": {
         "type": "git",
-        "url": "https://github.com/keldenl/fleece"
+        "url": "https://github.com/cafesao/fleece"
     },
     "engines": {
-        "vscode": "^1.0.0"
+        "vscode": "^1.86.0"
     },
     "categories": [
         "Other"
     ],
     "icon": "./logo-trans.png",
     "activationEvents": [
-        "*",
-        "onCommand:fleece.commentToCode",
-        "onCommand:fleece.startDalai",
-        "onCommand:fleece.stopFleece"
+        "*"
     ],
     "main": "./src/index.js",
     "contributes": {
-        "commands": [{
+        "commands": [
+            {
                 "command": "fleece.commentToCode",
                 "title": "Convert Comment to Code",
                 "category": "Fleece"
@@ -45,29 +43,47 @@
             }
         ],
         "menus": {
-            "editor/context": [{
-                "command": "fleece.commentToCode",
-                "group": "edit"
-            }]
+            "editor/context": [
+                {
+                    "command": "fleece.commentToCode",
+                    "group": "edit"
+                }
+            ]
         },
-        "keybindings": [{
-            "command": "fleece.commentToCode",
-            "key": "ctrl+alt+c",
-            "mac": "cmd+alt+c",
-            "when": "editorTextFocus && !editorReadonly"
-        }]
+        "keybindings": [
+            {
+                "command": "fleece.commentToCode",
+                "key": "ctrl+alt+c",
+                "mac": "cmd+alt+c",
+                "when": "editorTextFocus && !editorReadonly"
+            }
+        ]
     },
     "scripts": {
-        "make": "vsce package",
-        "deploy": "vsce publish"
+        "vscode:prepublish": "pnpm run compile",
+        "compile": "tsc -p ./",
+        "watch": "tsc -watch -p ./",
+        "pretest": "pnpm run compile && pnpm run lint",
+        "lint": "eslint src --ext ts",
+        "test": "vscode-test"
     },
     "devDependencies": {
-        "@types/vscode": "^1.0.0",
-        "typescript": "^5.0.2",
-        "vscode-test": "^1.6.1",
-        "@vscode/vsce": "^2.18.0"
+        "@types/mocha": "^10.0.6",
+        "@types/node": "20.11.21",
+        "@types/vscode": "^1.86.0",
+        "@typescript-eslint/eslint-plugin": "^7.1.0",
+        "@typescript-eslint/parser": "^7.1.0",
+        "@vscode/test-cli": "^0.0.6",
+        "@vscode/test-electron": "^2.3.9",
+        "eslint": "^8.57.0",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-prettier": "^5.1.3",
+        "prettier": "^3.2.5",
+        "source-map-support": "^0.5.21",
+        "typescript": "^5.3.3"
     },
     "dependencies": {
-        "socket.io-client": "^4.6.1"
+        "pino": "^8.19.0",
+        "socket.io-client": "^4.7.4"
     }
 }
